balancetracker-tray/requirements.txt
requests>=2.28.0
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
